fix: CommentsService 에서 댓글 삭제시 유저 확인 부분 추가   - 다른 사람이면 IllegalArgumentException throw

+++ 003 Code/BE/notion-linked-blog/src/main/java/io/f12/notionlinkedblog/service/comments/CommentsService.java
@@ -58,7 +58,7 @@ public class CommentsService {
 
 		return CommentSearchDto.builder()
 			.username(user.getUsername())
-			.comments(comments.getContent())
+			.comments(savedComments.getContent())
 			.build();
 	}
 
@@ -77,8 +77,11 @@ public class CommentsService {
 	}
 
 	public void removeComment(Long commentId, Long userId) {
-		commentsDataRepository.findById(commentId)
+		Comments comments = commentsDataRepository.findById(commentId)
 			.orElseThrow(() -> new IllegalArgumentException(COMMENT_NOT_EXIST));
+		if (!isSameUser(userId, comments.getUser().getId())) {
+			throw new IllegalStateException(NOT_COMMENT_OWNER);
+		}
 		commentsDataRepository.removeByIdAndUserId(commentId, userId);
 	}
